sufro

+++ P2_Grupo02/P2/Game/src/main/java/com/example/androidgame/GameLogic/ShopItem.java
@@ -9,14 +9,12 @@ public class ShopItem extends ButtonImage{
 
     private String sectionId_;
     private String itemId_;
-    private boolean unlocked_;
     private int price_;
     protected Image overlayImage;
 
     public ShopItem(String image, int w, int h, int x, int y, Sound buttonSound,
                      String sectionName,String itemName,int price) {
         super(image, w, h, x, y, buttonSound, null);
-        unlocked_=false;
         sectionId_=sectionName;
         itemId_=itemName;
         price_=price;
@@ -30,26 +28,21 @@ public class ShopItem extends ButtonImage{
         price_=price;
     }
 
-    public void buyItem(){
-
+    public boolean buyItem(){
+        boolean canBuy=false;
         int currCoins = GameManager.getInstance().getCoins();
-        unlocked_ = true;
-        ShopManager.getInstance().changeItemState(sectionId_, itemId_, unlocked_);
-        /*if(currCoins-price_>=0){
+        if(currCoins-price_>=0 && ShopManager.getInstance().itemIsLocked(sectionId_,itemId_)){
             GameManager.getInstance().addCoins(-price_);
-
+            ShopManager.getInstance().changeItemState(sectionId_,itemId_,true);
             deleteOverlayImage();
-        }*/
-
+            canBuy=true;
+        }
+        return canBuy;
     }
-    public boolean isUnlocked(){
-        return unlocked_;
-    }
-    public void setUnlocked(boolean isUnlocked){unlocked_=true;}
     @Override
     public void render(Graphics graph){
         super.render(graph);
-        if (!unlocked_)
+        if (ShopManager.getInstance().itemIsLocked(sectionId_,itemId_))
             graph.drawImage(overlayImage, this.posX_+width_/4, this.posY_+height_/4, width_/2, height_/2);
     }
     public void addOverlayImage(Image img) {
